rename index folder to match case
