+++ frontend/src/App.js
@@ -2,7 +2,7 @@ import React from 'react';
 import {
    BrowserRouter as Router, Switch, Route
 } from 'react-router-dom';
-import Index from './pages/index/Index';
+import Index from './pages/Index/Index';
 import Calendars from './pages/Calendars/Calendars';
 import Room from './pages/Room/Room';
 import PageTemplate from './templates/PageTemplate/PageTemplate';
